Reorganize test folder for consistency
