+++ src/peer_store.rs
@@ -173,7 +173,7 @@ impl TryFrom<String> for PeerInfo {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::tests::test_utils::TestPersister;
+	use crate::test::utils::TestPersister;
 	use proptest::prelude::*;
 	use std::str::FromStr;
