+++ src/test/mod.rs
@@ -1,2 +1,2 @@
 pub mod functional_tests;
-pub mod test_utils;
+pub mod utils;
